requirements.txt
pymongo
flask
flask_api
requests
pymysql
pyopenssl
